+++ jans-casa/plugins/samples/helloworld/src/main/java/io/jans/casa/plugins/helloworld/HelloWorldMenu.java
@@ -1,0 +1,26 @@
+package io.jans.casa.plugins.helloworld;
+
+import io.jans.casa.extension.navigation.MenuType;
+import io.jans.casa.extension.navigation.NavigationMenu;
+import org.pf4j.Extension;
+
+/**
+ * An extension class implementing the {@link NavigationMenu} extension point.
+ * @author jgomer
+ */
+@Extension
+public class HelloWorldMenu implements NavigationMenu {
+
+    public String getContentsUrl() {
+        return "menu.zul";
+    }
+
+    public MenuType menuType() {
+        return MenuType.USER;
+    }
+
+    public float getPriority() {
+        return 0.5f;
+    }
+
+}