+++ keycloak-integration/authenticator/src/main/java/io/jans/kc/spi/auth/oidc/OIDCMetaCacheKeys.java
@@ -1,0 +1,7 @@
+package io.jans.kc.spi.auth.oidc;
+
+public class OIDCMetaCacheKeys {
+    public static final String AUTHORIZATION_URL = "oidc.authorization.url";
+    public static final String TOKEN_URL = "oidc.token.url";  
+    public static final String USERINFO_URL = "oidc.userinfo.url";
+}