+++ keycloak-integration/authenticator/src/main/java/io/jans/kc/spi/auth/oidc/OIDCUserInfoRequestError.java
@@ -1,0 +1,13 @@
+package io.jans.kc.spi.auth.oidc;
+
+public class OIDCUserInfoRequestError extends Exception {
+    
+    public OIDCUserInfoRequestError(String msg) {
+        super(msg);
+    }
+
+    
+    public OIDCUserInfoRequestError(String msg, Throwable cause) {
+        super(msg,cause);
+    }
+}